requirements.txt
# C'est la liste des dépendances nécessaires
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
python-dotenv==0.19.2